Add files via upload
This is an exploratory data analysis assignment done in my Data Analytics and Big Data CS624 course taken summer 2025. This assignment was created and posed as an introduction to my use of Python with my partner Jasmine Armster. We worked on the code together and the added .csv compliments the code we used to complete the assignment. I added this 1st notebook I ever created because there were elements I hoped to update in the future as I've expanded my knowledge of Python.