Add files via upload
This uploaded jupyter notebook is a final project used in my CS620 class for fake news detection. It was the culmination of learning basic Python coding, GENAI, and ML concepts with my partner Nick Carpenter. The final assignment was accompanied with a final power point discussing our process and results.