Add files via upload
This Spotify_Tracks notebook was my first Python based assignment in my CS624 Data Analytics and Big Data course at ODU. Being new to Python it was an exploratory data analysis assignment to have us become familiar with the use of key pandas and big data function. The .csv file is too large to include in the repo but can be reviewed.